apply requested changes from review (add status field and rename app api to company)
