+++ backend/backend/urls.py
@@ -1,9 +1,9 @@
 from django.contrib import admin
 from django.urls import path, include
 
-from api.routers import router
+from company.routers import router as company_router
 
 urlpatterns = [
     path("admin/", admin.site.urls),
-    path('', include(router.urls)),
+    path('', include(company_router.urls)),
 ]
